add soft deletes article & tag, policy article

+++ app/Http/Controllers/Backend/ArticleController.php
@@ -10,6 +10,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\ArticleRequest;
 use App\Http\Services\Backend\ArticleService;
 use App\Http\Services\Backend\ImageService;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -67,8 +68,12 @@ class ArticleController extends Controller
      */
     public function show(string $uuid)
     {
+        $article = $this->articleService->getFirstBy('uuid', $uuid, true);
+
+        Gate::authorize('view', $article);
+
         return view('backend.articles.show', [
-            'article' => $this->articleService->getFirstBy('uuid', $uuid, true),
+            'article' => $article,
         ]);
     }
 
@@ -77,8 +82,12 @@ class ArticleController extends Controller
      */
     public function edit(string $uuid)
     {
+        $article = $this->articleService->getFirstBy('uuid', $uuid, true);
+
+        Gate::authorize('view', $article);
+
         return view('backend.articles.edit', [
-            'article' => $this->articleService->getFirstBy('uuid', $uuid, true),
+            'article' => $article,
             'categories' => $this->articleService->getCategory(),
             'tags' => $this->articleService->getTag()
         ]);
@@ -117,9 +126,37 @@ class ArticleController extends Controller
      */
     public function destroy(string $uuid)
     {
+        $article = $this->articleService->getFirstBy('uuid', $uuid, true);
+
+        Gate::authorize('view', $article);
+
         $this->articleService->delete($uuid);
 
         return response()->json(['message' => 'Data Artikel Berhasil Dihapus...']);
+    }
+
+    public function forceDelete(string $uuid)
+    {
+        $article = $this->articleService->getFirstBy('uuid', $uuid, true);
+
+        Gate::authorize('view', $article);
+
+        $this->articleService->forceDelete($uuid);
+
+        return response()->json([
+            'message' => 'Data Artikel Berhasil Dihapus Permanen...',
+        ]);
+    }
+
+    public function restore(string $uuid)
+    {
+        $article = $this->articleService->getFirstBy('uuid', $uuid, true);
+
+        Gate::authorize('view', $article);
+
+        $this->articleService->restore($uuid);
+
+        return redirect()->back()->with('success', 'Data Artikel Berhasil Dipulihkan...');
     }
 
     public function serverside(Request $request): JsonResponse
